requirements.txt
gpiozero
picamera
requests
pillow
firebase-admin
gradio_client